Move search features to nav bar and restrict search display to root page.
